Split interface into multiple files
